Refactored To make interface lay with it's implementation
